Corrections made successfully to the JPA configuration

+++ src/main/java/br/com/cadproduto/ws/repository/ClienteRepository.java
@@ -8,7 +8,7 @@ import org.springframework.transaction.annotation.Transactional;
 import br.com.cadproduto.ws.model.Produto;
 @Repository
 @Transactional
-public interface ClienteRepository extends JpaRepository<Produto, Integer>{
+public interface ClienteRepository extends JpaRepository<Produto, Long>{
 	
 }
